roll out grid_sampler_v2
